Compute node left width based on widest control.

+++ com.unity.visualeffectgraph/VisualEffectGraph/Editor/RMGUI/Vector3Field.cs
@@ -7,9 +7,9 @@ namespace UnityEditor.VFX.UIElements
 {
     class Vector3Field : VFXControl<Vector3>
     {
-        LabeledField<DoubleField, double> m_X;
-        LabeledField<DoubleField, double> m_Y;
-        LabeledField<DoubleField, double> m_Z;
+        LabeledField<FloatField, float> m_X;
+        LabeledField<FloatField, float> m_Y;
+        LabeledField<FloatField, float> m_Z;
 
         public bool dynamicUpdate
         {
@@ -26,9 +26,9 @@ namespace UnityEditor.VFX.UIElements
         }
         void CreateTextField()
         {
-            m_X = new LabeledField<DoubleField, double>("X");
-            m_Y = new LabeledField<DoubleField, double>("Y");
-            m_Z = new LabeledField<DoubleField, double>("Z");
+            m_X = new LabeledField<FloatField, float>("X");
+            m_Y = new LabeledField<FloatField, float>("Y");
+            m_Z = new LabeledField<FloatField, float>("Z");
 
             m_X.control.AddToClassList("fieldContainer");
             m_Y.control.AddToClassList("fieldContainer");
@@ -37,26 +37,26 @@ namespace UnityEditor.VFX.UIElements
             m_Y.AddToClassList("fieldContainer");
             m_Z.AddToClassList("fieldContainer");
 
-            m_X.RegisterCallback<ChangeEvent<double>>(OnXValueChanged);
-            m_Y.RegisterCallback<ChangeEvent<double>>(OnYValueChanged);
-            m_Z.RegisterCallback<ChangeEvent<double>>(OnZValueChanged);
+            m_X.RegisterCallback<ChangeEvent<float>>(OnXValueChanged);
+            m_Y.RegisterCallback<ChangeEvent<float>>(OnYValueChanged);
+            m_Z.RegisterCallback<ChangeEvent<float>>(OnZValueChanged);
         }
 
-        void OnXValueChanged(ChangeEvent<double> e)
+        void OnXValueChanged(ChangeEvent<float> e)
         {
             Vector3 newValue = value;
             newValue.x = (float)m_X.value;
             SetValueAndNotify(newValue);
         }
 
-        void OnYValueChanged(ChangeEvent<double> e)
+        void OnYValueChanged(ChangeEvent<float> e)
         {
             Vector3 newValue = value;
             newValue.y = (float)m_Y.value;
             SetValueAndNotify(newValue);
         }
 
-        void OnZValueChanged(ChangeEvent<double> e)
+        void OnZValueChanged(ChangeEvent<float> e)
         {
             Vector3 newValue = value;
             newValue.z = (float)m_Z.value;
